tests(app): add unit tests for Database, move tests into its own file
